+++ FinacPOS/Transaction/frmPOSSales2.cs
@@ -1,4 +1,5 @@
-﻿using DevExpress.XtraRichEdit.Model;
+﻿using DevExpress.XtraReports.UI;
+using DevExpress.XtraRichEdit.Model;
 using FinacPOS.Masters;
 using Microsoft.VisualBasic;
 using OnBarcode.Barcode.WinForms;
@@ -28,10 +29,11 @@ namespace FinacPOS
         }
         public void setLanguage(String language)
         {
+
             //Thread.CurrentThread.CurrentUICulture = new System.Globalization.CultureInfo(language);
             if (language == "ARB")
             {
-                this.RightToLeft = RightToLeft.Yes;
+                this.RightToLeft = System.Windows.Forms.RightToLeft.Yes;
                 this.RightToLeftLayout = true;
             }
             //this.Controls.Clear();
@@ -59,7 +61,7 @@ namespace FinacPOS
         string strGridValue = "";
         int CurColIndex = 0;
         int CurEditRowIndex = 0;
-
+        public string selectedSalesMode = "";
         //varis form thermal print
         DataTable dtblCompanyDetailsThermal;
         DataTable dtblGridDetailsThermal;
@@ -80,7 +82,7 @@ namespace FinacPOS
         ProductSP spProduct = new ProductSP();
 
         DataGridViewTextBoxEditingControl TextBoxControl;
-
+        POSSettingsInfo InfoPOSSettings = new POSSettingsInfo();
         //for Posting Account setting
         string strCashSalesLedgerId = "";
         string strCCSalesLedgerId = "";
@@ -90,6 +92,8 @@ namespace FinacPOS
         bool IsAuthenticationApproved = false;
 
         DataTable dtblProductWithImage;
+        string strBarcode = "";
+       int strDeleteCurrentRowIndex = 0;
         #endregion
 
         #region FUNCTIONS
@@ -1161,7 +1165,7 @@ namespace FinacPOS
             BranchSP SpBranch = new BranchSP();
             dtblCompanyDetails = SpBranch.BranchViewByBranchId(PublicVariables._branchId);
             POSSalesMasterSP salesmaster = new POSSalesMasterSP();
-
+            POSSalesMasterInfo InfoPOSSalesMaster = new POSSalesMasterInfo();
             dtblCompanyDetails.Columns.Add("companyheader_logo");
             //dtblCompanyDetails.Columns.Add("companyfooter_logo");
 
@@ -1320,6 +1324,7 @@ namespace FinacPOS
             dtblOtherDetails.Columns.Add("CustomerPhone");
             dtblOtherDetails.Columns.Add("CustomerVatNo");
             dtblOtherDetails.Columns.Add("TokenNo");
+            dtblOtherDetails.Columns.Add("SalesMode");
 
             if (isDuplicatePrint == false)
             {
@@ -1436,6 +1441,7 @@ namespace FinacPOS
                     dRowDetails["BillAmount"] = "";
                     dRowDetails["totalBalance"] = "";
                     dRowDetails["showCustBalance"] = false;
+                    dRowDetails["SalesMode"] = selectedSalesMode;
                 }
 
                 //////------------------------ QR Code Generation ----------- by Navas --------------------
@@ -1622,7 +1628,7 @@ namespace FinacPOS
                     }
 
                     dRowDetails["BillAmount"] = Convert.ToDecimal(dtbl.Rows[0]["totalAmount"]).ToString(FinanceSettingsInfo._roundDecimalPart);
-
+                    dRowDetails["SalesMode"] = dtbl.Rows[0]["SalesMode"].ToString();
 
                     //////------------------------ QR Code Generation ----------- by Navas --------------------
                     ////Zen.Barcode.CodeQrBarcodeDraw qrBarcode = Zen.Barcode.BarcodeDrawFactory.CodeQr;
@@ -2360,20 +2366,24 @@ namespace FinacPOS
             InfoPOSSalesMaster.CreditNoteNo = strCreditNoteNo;
             InfoPOSSalesMaster.CreditNoteAmount = decCreditNoteAmt;
             InfoPOSSalesMaster.UserId = PublicVariables._currentUserId;
+            
+
             if (rbDineIn.Checked == true) //ADDED on 29-03-2025 by nishana
             {
-                  InfoPOSSalesMaster.SalesMode =  rbDineIn.Text;
+                InfoPOSSalesMaster.SalesMode = rbDineIn.Text;
+                selectedSalesMode = rbDineIn.Text;
             }
             else if (rbTakeway.Checked == true)
             {
                 InfoPOSSalesMaster.SalesMode = rbTakeway.Text;
+                selectedSalesMode = rbTakeway.Text;
             }
             else if (rbOnline.Checked == true)
             {
                 InfoPOSSalesMaster.SalesMode = rbOnline.Text;
-
+                selectedSalesMode = rbOnline.Text;
             }
-           
+
             InfoPOSSalesMaster.CustomerAddress = "";
             InfoPOSSalesMaster.CustomerPhone = "";
             InfoPOSSalesMaster.CustomerVATNo = "";
@@ -3042,56 +3052,62 @@ namespace FinacPOS
                 lblBarcodeScanningType.Visible = false; 
                 return;
             }
-            // Delete Product from Grid 
-             bool IsBarcodeExist = false; // added on 12-04-2025 by Nishana
-            int RoWnO = -1;
-            string Barcode = txtBarcode.Text.Trim();
 
-            if (lblBarcodeScanningType.Visible == true && lblBarcodeScanningType.Text == "Delete Product")
+
+            // Delete Product from Grid when barcode is entered
+            if (InfoPOSSettings.DeleteMode == "Delete By Barcode")
             {
+                bool IsBarcodeExist = false; // added on 12-04-2025 by Nishana
+                int RoWnO = -1;
+                string Barcode = txtBarcode.Text.Trim();
 
-                for (int i = 0; i < dgvProduct.RowCount; i++)
+                if (lblBarcodeScanningType.Visible == true && lblBarcodeScanningType.Text == "Delete Product")
                 {
-                    if (dgvProduct.Rows[i].Cells["Barcode"].Value != null && dgvProduct.Rows[i].Cells["Barcode"].Value.ToString() == Barcode)
+
+                    for (int i = 0; i < dgvProduct.RowCount; i++)
                     {
-                        IsBarcodeExist = true;
-                        RoWnO = i;
-                        break;
+                        if (dgvProduct.Rows[i].Cells["Barcode"].Value != null && dgvProduct.Rows[i].Cells["Barcode"].Value.ToString() == Barcode)
+                        {
+                            IsBarcodeExist = true;
+                            RoWnO = i;
+                            break;
 
+                        }
                     }
-                }
-                if (IsBarcodeExist == true)
-                {
-                    if (!dgvProduct.Rows[RoWnO].IsNewRow)
+                    if (IsBarcodeExist == true)
                     {
-                        int selectedRowIndex = dgvProduct.CurrentCell.RowIndex;
-                        SaveGridDeletedItem(selectedRowIndex);
-
-            
-                        dgvProduct.Rows.RemoveAt(RoWnO);
-
-
-                        if (dgvProduct.Rows.Count > 1)
+                        if (!dgvProduct.Rows[RoWnO].IsNewRow)
                         {
-                            dgvCurRow = dgvProduct.Rows.Count - 1;
-                        }
-                        else
-                        {
-                            dgvCurRow = 0;
-                            dgvSlno = 1;
-                        }
+                            int selectedRowIndex = dgvProduct.CurrentCell.RowIndex;
+                            SaveGridDeletedItem(selectedRowIndex);
 
-                        CalculateBillDiscforIndivProduct();
-                        CalculateBillTotal();
+
+                            dgvProduct.Rows.RemoveAt(RoWnO);
+
+
+                            if (dgvProduct.Rows.Count > 1)
+                            {
+                                dgvCurRow = dgvProduct.Rows.Count - 1;
+                            }
+                            else
+                            {
+                                dgvCurRow = 0;
+                                dgvSlno = 1;
+                            }
+
+                            CalculateBillDiscforIndivProduct();
+                            CalculateBillTotal();
+                        }
                     }
-                }
 
-                lblBarcodeScanningType.Text = "";
-                lblBarcodeScanningType.Visible = false;
-              
-                txtBarcode.Text = "";
-                return;
-            }                     
+                    lblBarcodeScanningType.Text = "";
+                    lblBarcodeScanningType.Visible = false;
+
+                    txtBarcode.Text = "";
+                    return;
+                }
+            }
+                          
 
             DataTable dtbl = new DataTable();
             dtbl = SPGeneral.GetProductDetailsByBarcode(txtBarcode.Text.Trim());
@@ -4599,34 +4615,40 @@ namespace FinacPOS
 
             
             barcodeFocus();
-            //if (dgvProduct.RowCount > 1)
-            //{
-            //    if (!dgvProduct.CurrentRow.IsNewRow)
-            //    {
-            //        DataGridViewRow CurRow = dgvProduct.CurrentRow;
-            //        dgvProduct.Rows.Remove(CurRow);
+            if (InfoPOSSettings.DeleteMode == "Delete By Button Click")
+            {
 
-            //        if (dgvCurRow > 1)
-            //        {
-            //            dgvCurRow = dgvProduct.Rows.Count - 1;
-            //        }
-            //        else
-            //        {
-            //            dgvCurRow = 0;
-            //            dgvSlno = 1;
-            //        }
-                    
+                if (dgvProduct.RowCount > 1)
+                {
+                    if (!dgvProduct.CurrentRow.IsNewRow)
+                    {
 
-            //        CalculateBillDiscforIndivProduct();
-            //        CalculateBillTotal();
-            //    }
+                        int selectedRowIndex = strDeleteCurrentRowIndex;
+                        SaveGridDeletedItem(selectedRowIndex);
+                        DeleteRow(selectedRowIndex);
 
-            //    barcodeFocus();
-            //}
-            //else
-            //{
-            //    barcodeFocus();
-            //}
+                        if (dgvCurRow > 1)
+                        {
+                            dgvCurRow = dgvProduct.Rows.Count - 1;
+                        }
+                        else
+                        {
+                            dgvCurRow = 0;
+                            dgvSlno = 1;
+                        }
+
+                        CalculateBillDiscforIndivProduct();
+                        CalculateBillTotal();
+                    }
+                    barcodeFocus();
+                }
+
+                else
+                {
+                    barcodeFocus();
+                }
+            }
+
         }
         //string strButtonNameCondition = "";
         private void btnCash_Click(object sender, EventArgs e)
@@ -4782,37 +4804,42 @@ namespace FinacPOS
             }
             if (e.KeyCode == Keys.Delete)
             {
-                if (dgvProduct.RowCount > 1)
+                if (InfoPOSSettings.DeleteMode == "Delete By Keyboard Key")
                 {
-                    if (!dgvProduct.CurrentRow.IsNewRow)
+                    if (dgvProduct.RowCount > 1)
                     {
+                        if (!dgvProduct.CurrentRow.IsNewRow)
                         {
-                            int selectedRowIndex = dgvProduct.CurrentCell.RowIndex;
-                            SaveGridDeletedItem(selectedRowIndex);
-
-                            DeleteRow(selectedRowIndex);
-
-                            if (dgvCurRow > 1)
                             {
-                                dgvCurRow = dgvProduct.Rows.Count - 1;
-                            }
-                            else
-                            {
-                                dgvCurRow = 0;
-                                dgvSlno = 1;
+                                int selectedRowIndex = dgvProduct.CurrentCell.RowIndex;
+                                SaveGridDeletedItem(selectedRowIndex);
+
+                                DeleteRow(selectedRowIndex);
+
+                                if (dgvCurRow > 1)
+                                {
+                                    dgvCurRow = dgvProduct.Rows.Count - 1;
+                                }
+                                else
+                                {
+                                    dgvCurRow = 0;
+                                    dgvSlno = 1;
+                                }
+
+                                CalculateBillDiscforIndivProduct();
+                                CalculateBillTotal();
                             }
 
-                            CalculateBillDiscforIndivProduct();
-                            CalculateBillTotal();
+                            barcodeFocus();
                         }
+                        else
+                        {
+                            barcodeFocus();
+                        }
+                    }
 
-                        barcodeFocus();
-                    }
-                    else
-                    {
-                        barcodeFocus();
-                    }
                 }
+                   
             }
         }
 
@@ -5512,6 +5539,20 @@ namespace FinacPOS
                 dgvProduct.Rows[e.RowIndex].Cells[e.ColumnIndex].ToolTipText = "PR: " + purchaseRate.ToString() + " (Excl. Tax), " + IncludeRate.ToString() + " (Incl. Tax), Stock: " + Stock.ToString();
 
 
+            }
+        }
+
+        private void dgvProduct_CellClick(object sender, DataGridViewCellEventArgs e)
+        {
+
+            if (e.RowIndex >= 0)
+            {
+                try
+                {
+                    strBarcode = dgvProduct.Rows[e.RowIndex].Cells["Barcode"].Value.ToString();
+                    strDeleteCurrentRowIndex = e.RowIndex;
+                }
+                catch { }
             }
         }
     }
